requirements.txt
fastapi==0.104.0
gunicorn==21.2.0
httpx==0.25.0
numpy==1.24.3
pandas==2.0.3
pytest==7.4.0
requests==2.31.0
scikit-learn==1.3.0
uvicorn==0.23.2
omegaconf==2.3.0
hydra-core==1.3.2
pydantic==1.10.12